Fix function body resolution
When I switched up function body resolution in order to handle recursive
functions, I introduced a regression. The function was being added to
the ordered symbols list before the body was resolved, so any
declarations that statements in the function depended on, were getting
added to ordered symbols after the function itself. This patch fixes
that regression.