Fix resolution of function signatures and function bodies
Previously I was resolving function bodies any time the function symbol
was resolved. So for example, resolving a call expression would trigger
the whole function body resolution. This led to issues like the local
symbol table having variables in scope that should not be in the parent
scope of the called function. The crux of the issue was not
differentiating between resolving a function definition vs any other
reference to that function. Now, function bodies are only resolved when
the function symbol is encountered in the context of the function
definition.

Care had to be taken also to ensure that the function symbol is not
added to the ordered symbol table until after the function body has been
resolved.